window-list: prevent infinite loop when too many windows open.
Currently, when the list becomes larger than the maximum size,
on_draw calls set_button_width to recalculate the button sizes.
This normally results in the buttons getting smaller. But if the
buttons are already at the minimum size, this results in an
infinite loop where set_button_width allocates the same width
as it did before, and on_draw calls it again and again.

Fix this by caching the last button size and only redrawing if it
changed.